requirements.txt
readability-lxml
html2text